resolve all errors
