+++ package.json
@@ -100,6 +100,7 @@
     "eslint-plugin-unused-imports": "^3.2.0",
     "jest": "^29.7.0",
     "jest-watch-typeahead": "^2.2.2",
+    "madge": "^7.0.0",
     "next": "13.4.13",
     "nock": "^13.5.4",
     "prettier": "^3.3.3",
