+++ src/util/json.ts
@@ -1,0 +1,25 @@
+export function isValidJson(str: string): boolean {
+  try {
+    JSON.parse(str);
+    return true;
+  } catch (e) {
+    return false;
+  }
+}
+
+export function safeJsonStringify(value: any, prettyPrint: boolean = false): string {
+  // Prevent circular references
+  const cache = new Set();
+  const space = prettyPrint ? 2 : undefined;
+  return JSON.stringify(
+    value,
+    (key, val) => {
+      if (typeof val === 'object' && val !== null) {
+        if (cache.has(val)) return;
+        cache.add(val);
+      }
+      return val;
+    },
+    space,
+  );
+}
